+++ composeApp/src/commonMain/kotlin/org/wip/plugintoolkit/features/job/model/Job.kt
@@ -20,6 +20,7 @@ enum class JobType {
     Capability,
     Flow,
     Setup,
+    Update,
     Validation,
     PluginAction
 }